pyproject.toml

[project]
name = ""
version = ""
description = ""
authors = [
    {name = "horodeski", email = "geovana.horodeski06@gmail.com"},
]
dependencies = [
    "django>=4.2.2",
    "djangorestframework>=3.14.0",
    "django-cors-headers>=4.1.0",
    "djangorestframework-simplejwt>=5.2.2",
]
requires-python = ">=3.10"
license = {text = "MIT"}

[tool.pdm.dev-dependencies]
dev = [
    "black>=23.3.0",
]
